requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
kubernetes==28.1.0
requests==2.31.0
PyYAML==6.0.1
Werkzeug==2.3.7 